// Frontend.react/scpex-front/src/axios/models/user.ts
// tslint:disable
/**
 * SCPExC
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface User
 */
export interface User {
    /**
     * 
     * @type {number}
     * @memberof User
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    contact: string;
    /**
     * 
     * @type {number}
     * @memberof User
     */
    type: number;
    /**
     * 
     * @type {number}
     * @memberof User
     */
    status: number;
    /**
     * 
     * @type {number}
     * @memberof User
     */
    createdDate: number;
    /**
     * 
     * @type {number}
     * @memberof User
     */
    birth: number;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    passwd: string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    note: string;
}